major code refactoring to separate npu-driven ha scope and dpu-driven ha scope
